src/requirements.txt
functions-framework==3.5.0
google-genai==1.70.0
google-cloud-storage==2.14.0
google-cloud-texttospeech==2.16.3
requests==2.33.1
